[UT] Move agg_hash_map benchmark to be/src/bench
Extract performance benchmarks from agg_hash_map_bench_test.cpp into
a dedicated Google Benchmark binary (be/src/bench/agg_hash_map_bench.cpp),
converting chrono-based timing to benchmark::State loops. Keep only the
correctness test (InlineStateCorrectness) in be/test as
agg_hash_map_inline_test.cpp.
