+++ be/src/bench/CMakeLists.txt
@@ -40,3 +40,4 @@ ADD_BE_BENCH(${SRC_DIR}/bench/allocator_bench)
 ADD_BE_BENCH(${SRC_DIR}/bench/jemalloc_bench)
 ADD_BE_BENCH(${SRC_DIR}/bench/buffer_bench)
 ADD_BE_BENCH(${SRC_DIR}/bench/lock_free_scan_task_queue_bench)
+ADD_BE_BENCH(${SRC_DIR}/bench/agg_hash_map_bench)
